Major reorganisation of the solution
AlgoLib now depends on {CSDiaballik => DiaballikCore}, which allows the
library to inspect the gameboard without having to awkwardly copy state
into a C++ object. Since the AiPlayers depended on the C++ library,
their implementation is moved out of DiaballikCore into a new
DiaballikPlayers assembly. The interface IPlayer is left in Core to
bridge the gap.

Namespaces for all the assemblies have been updated to a more sensible
convention.

The dependency graph is now the following:
DiaballikCore <- AlgoLib <- DiaballikPlayers <- DiaballikTests -> DiaballikWPF
         \__________\________________________<_________|