Remove GetNextMove from the player's interface.
Players are now pure data holders, which simplifies their
use throughout the project. They hold their type as an attribute.
We don't need a class hierarchy anymore, a single class is enough.

This change was prompted by the need to treat human players very
differently from AI players. The behaviour of AI players is still
held in an AiDecisionAlgo. ProgressiveAiAlgo is moved to the C++ lib.

This removes the need for the DiaballikPlayers DLL. The WPF project
can use the AiDecisionAlgos from AlgoLib.dll and the rest of the classes
is provided by DiaballikCore.dll.

Player builders no longer need a non-blank name to build. We'll use
a generated name for players who don't chose their names.