Implement save / load logic.
The core logic is located in SaveManager.cs.

Games have an identity (represented by an id), which
is preserved throughout the load / save cycle and until the deletion of the save.

Forking a game creates a game with a new id. Playing actions on a game applies the
changes with the same id, thus overwriting previous saves.

Each memento is contained in its own file. One manifest file gathers metadata for all
known saves. That information is lazy loaded and its metadata can be displayed in the load
screen's itemscontrol. When a save is selected, the corresponding memento file is fetched
and parsed, and sent to the game view. Laziness everywhere keeps disk operations at a minimum.